Rename webpack package loader property with resolve loader and update version
