+++ packages/webpack/package.json
@@ -1,6 +1,6 @@
 {
   "name": "build-tools-webpack",
-  "version": "1.1.2",
+  "version": "1.2.0",
   "description": "Configurations to use webpack",
   "main": "./index.js",
   "license": "MIT",
